Change entry point to method instead of class
